# backend/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/article:
    get:
      parameters:
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      - description: uid
        in: query
        name: uid
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 获取文章列表
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: integer
      - description: 内容
        in: formData
        name: content
        required: true
        type: string
      - description: 标签
        in: formData
        items:
          type: integer
        name: tag_name
        type: array
      - description: image
        in: formData
        name: images
        type: file
      - description: video
        in: formData
        name: video
        type: file
      - description: 视频类型为1
        in: formData
        name: type
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 添加文章
  /api/v1/article/{id}:
    delete:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 删除文章
    put:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: title
        in: formData
        name: title
        type: string
      - description: content
        in: formData
        name: content
        type: string
      produces:
      - application/json
      responses: {}
      summary: 更新文章
  /api/v1/article/{id}/addTags:
    post:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签
        in: formData
        items:
          type: string
        name: tag_name
        type: array
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 添加文章标签
  /api/v1/article/{id}/deleteTags:
    delete:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签
        in: formData
        items:
          type: string
        name: tag_name
        required: true
        type: array
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 删除文章标签
  /api/v1/article/{id}/like:
    post:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户ID
        in: formData
        name: user_id
        required: true
        type: integer
      - description: 类型
        in: formData
        name: type
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 添加文章点赞
  /api/v1/article/{id}/recover:
    put:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 恢复文章
...
swagger: "2.0"